End? should try to give high interval between cars

+++ SystemTimer.java
@@ -2,7 +2,7 @@ public class SystemTimer {
     final long startTime = System.nanoTime();
 
 
-    public int currentSeconds() {
-        return (int) (System.nanoTime() - startTime / 1000000000);
+    public double currentSeconds() {
+        return  ((System.nanoTime() - startTime) / 1000000000);
     }
 }
